Avanzando en el grafico de ppm
Aun no lo tengo listo, pero trato de hacer un grafico que muestre las ppm de un gas.
También inclui una enumeración para poder seleccionar el tipo de gas a mostrar.
A futuro hay que colocar lafunción de calibración en su propio menu.

+++ src/main.h
@@ -70,9 +70,9 @@ Remarks: By using the slope and a point of the line. The x(logarithmic value of 
          logarithmic coordinate, power of 10 is used to convert the result to non-logarithmic 
          value.
 **********************************************************************************/ 
-inline double MQGetPercentage(float rs_ro_ratio, const float *pcurve)
+inline float MQGetPercentage(float rs_ro_ratio, const float *pcurve)
 {
-  return (pow(10,( ((log(rs_ro_ratio)-pcurve[1])/pcurve[2]) + pcurve[0])));  
+  return (float)(pow(10,( ((log(rs_ro_ratio)-pcurve[1])/pcurve[2]) + pcurve[0])));  
 }
 
 
@@ -91,5 +91,4 @@ inline float MQRead()
   if(_dataIndex >= DATA_LEN) return -1; //fuera de los limites
   return MQResistanceCalculation(data[_dataIndex]);  
 }
-
 #endif
